V2.54: try not to disturb the running schedule script
If previous versions, if you start RPi manually while schedule script
is running, you may disturb the script as the “runScript.sh” script
will re-schedule the next shutdown and next startup, which may shift
the script step forward. Since this version, if the schedule script is
running and the scheduled startup is in the future, the “runScript.sh”
script will not be executed and the schedule script will not be
disturbed as long as you shutdown Raspberry Pi before the schedule
startup time.